ءؤ/45/ea-v3-fc-v3-fifa-v3-main/FC26_sale_coins_Bot/requirements.txt
# FC26 Gaming Bot Dependencies
# مكتبات بوت FC26 للألعاب

# Telegram Bot Library (Main requirement)
python-telegram-bot>=22.0

# Environment Variables
python-dotenv

# Optional: For enhanced features if needed
# requests>=2.28.0  # For HTTP requests if needed
# aiohttp>=3.8.0    # For async HTTP if needed
# cryptography       # For encryption if needed

# Note: The following are Python built-ins and don't need installation:
# sqlite3, asyncio, logging, re, datetime, typing, os, platform, sys